+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Add `From<ChunkKeyEncodingTraits>` for `ChunkKeyEncoding`
+
 ## [0.7.2] - 2023-12-17
 
 ### Added